+++ Phantasma.Business/src/Phantasma.Business.csproj
@@ -11,6 +11,7 @@
     <Version>1.0.8</Version>
   </PropertyGroup>
   <ItemGroup>
+    <PackageReference Include="BouncyCastle.NetCore" Version="1.9.0" />
     <PackageReference Include="Nethereum.Web3" Version="4.14.0" />
     <PackageReference Include="Serilog" Version="2.11.0" />
   </ItemGroup>
